+++ app/controllers/pg_hero/home_controller.rb
@@ -17,6 +17,7 @@ module PgHero
       @table_hit_rate = PgHero.table_hit_rate
       @missing_indexes = PgHero.missing_indexes
       @unused_indexes = PgHero.unused_indexes.select { |q| q["index_scans"].to_i == 0 }
+      @invalid_indexes = PgHero.invalid_indexes
       @good_cache_rate = @table_hit_rate >= 0.99 && @index_hit_rate >= 0.99
       @query_stats_available = PgHero.query_stats_available?
       @total_connections = PgHero.total_connections